move core library to `@fartlabs/rt`
