--- components/mod.ts
@@ -1,4 +1,0 @@
-export * from "./components.tsx";
-
-import * as components from "./components.tsx";
-components satisfies components.ComponentsInterface;
