+++ deno.jsonc
@@ -8,6 +8,7 @@
   },
   "imports": {
     "@fartlabs/jsonx": "jsr:@fartlabs/jsonx@^0.0.10",
+    "@fartlabs/rt": "jsr:@fartlabs/rt@^0.0.1",
     "rtx/": "./"
   },
   "compilerOptions": {
