+++ mod.ts
@@ -1,7 +1,12 @@
 /**
  * @module
  *
- * rtx is a minimalistic HTTP router library based on the `URLPattern` API.
+ * rtx is a minimalistic HTTP router library based on the `URLPattern` API in
+ * JSX.
  */
+
+import * as rtx from "./rtx.ts";
+
+rtx satisfies rtx.ComponentsInterface;
 
 export * from "./rtx.ts";
